Refactor CurrentDate and Sales
