+++ src/main/java/seedu/cafectrl/CafeCtrl.java
@@ -1,9 +1,7 @@
 package seedu.cafectrl;
 
 import seedu.cafectrl.command.Command;
-import seedu.cafectrl.data.Menu;
-import seedu.cafectrl.data.OrderList;
-import seedu.cafectrl.data.Pantry;
+import seedu.cafectrl.data.*;
 import seedu.cafectrl.parser.Parser;
 import seedu.cafectrl.ui.Ui;
